Fixed Buggy behavior in CreateAnyGiftPokemon Code
Confirmed buggy behavior from a small line optimization I made in this code a while back.

Was already running slower and showing signs of run off, but confirmed from a user case that this was in fact the case.

Added another line of 0x00000000
for safe measure to ensure correct exit after calling this through NPC